issue #115 rename : model -> remote/model/detail_plan
