--- client/app/src/main/java/com/example/witt/data/model/plan/detail_plan/request/MemoRequest.kt
@@ -1,5 +1,0 @@
-package com.example.witt.data.model.plan.detail_plan.request
-
-data class MemoRequest (
-    val Content: String
-    )
