+++ client/app/src/main/java/com/example/witt/data/model/remote/detail_plan/memo/EditMemoResponse.kt
@@ -1,6 +1,6 @@
-package com.example.witt.data.model.plan.memo
+package com.example.witt.data.model.remote.detail_plan.memo
 
-import com.example.witt.domain.model.plan.memo.EditMemoModel
+import com.example.witt.domain.model.remote.detail_plan.memo.EditMemoModel
 
 
 data class EditMemoResponse (
