+++ client/app/src/main/java/com/example/witt/data/model/remote/detail_plan/place/request/MemoRequest.kt
@@ -1,0 +1,5 @@
+package com.example.witt.data.model.remote.detail_plan.place.request
+
+data class MemoRequest (
+    val Content: String
+    )
