+++ client/app/src/main/java/com/example/witt/data/model/remote/detail_plan/search/flight/AddFlightResponse.kt
@@ -1,4 +1,4 @@
-package com.example.witt.data.model.search.flight
+package com.example.witt.data.model.remote.detail_plan.search.flight
 
 data class AddFlightResponse (
     val status:Boolean,
