+++ client/app/src/main/java/com/example/witt/data/repository/DetailPlanRepositoryImpl.kt
@@ -1,16 +1,16 @@
 package com.example.witt.data.repository
 
-import com.example.witt.data.model.plan.detail_plan.response.toAddPlaceResponseModel
-import com.example.witt.data.model.plan.memo.toEditMemoModel
-import com.example.witt.data.model.plan.memo.toMakeMemoModel
-import com.example.witt.data.source.remote.memo.EditMemoDataSource
-import com.example.witt.data.source.remote.memo.MakeMemoDataSource
-import com.example.witt.data.source.remote.plan.detail_plan.AddPlaceDataSource
-import com.example.witt.domain.model.plan.detail_plan.AddPlaceModel
-import com.example.witt.domain.model.plan.detail_plan.AddPlaceResponseModel
-import com.example.witt.domain.model.plan.detail_plan.toAddPlaceRequest
-import com.example.witt.domain.model.plan.memo.EditMemoModel
-import com.example.witt.domain.model.plan.memo.MakeMemoModel
+import com.example.witt.data.model.remote.detail_plan.place.response.toAddPlaceResponseModel
+import com.example.witt.data.model.remote.detail_plan.memo.toEditMemoModel
+import com.example.witt.data.model.remote.detail_plan.memo.toMakeMemoModel
+import com.example.witt.data.source.remote.detail_plan.memo.EditMemoDataSource
+import com.example.witt.data.source.remote.detail_plan.memo.MakeMemoDataSource
+import com.example.witt.data.source.remote.detail_plan.place.AddPlaceDataSource
+import com.example.witt.domain.model.remote.detail_plan.place.AddPlaceModel
+import com.example.witt.domain.model.remote.detail_plan.place.AddPlaceResponseModel
+import com.example.witt.domain.model.remote.detail_plan.place.toAddPlaceRequest
+import com.example.witt.domain.model.remote.detail_plan.memo.EditMemoModel
+import com.example.witt.domain.model.remote.detail_plan.memo.MakeMemoModel
 import com.example.witt.domain.repository.DetailPlanRepository
 import javax.inject.Inject
