+++ client/app/src/main/java/com/example/witt/data/source/remote/detail_plan/place/AddPlaceDataSourceImpl.kt
@@ -1,8 +1,8 @@
-package com.example.witt.data.source.remote.plan.detail_plan
+package com.example.witt.data.source.remote.detail_plan.place
 
 import com.example.witt.data.api.DetailPlanService
-import com.example.witt.data.model.plan.detail_plan.request.AddPlaceRequest
-import com.example.witt.data.model.plan.detail_plan.response.AddPlaceResponse
+import com.example.witt.data.model.remote.detail_plan.place.request.AddPlaceRequest
+import com.example.witt.data.model.remote.detail_plan.place.response.AddPlaceResponse
 import com.example.witt.di.IoDispatcher
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
@@ -11,7 +11,7 @@ import javax.inject.Inject
 class AddPlaceDataSourceImpl @Inject constructor(
     private val service : DetailPlanService,
     @IoDispatcher private val coroutineDispatcher: CoroutineDispatcher
-): AddPlaceDataSource{
+): AddPlaceDataSource {
 
     override suspend fun addPlace(tripId: Int, dayId: Int, request: AddPlaceRequest): Result<AddPlaceResponse>
     = withContext(coroutineDispatcher){
