--- client/app/src/main/java/com/example/witt/data/source/remote/plan/detail_plan/AddPlaceDataSource.kt
@@ -1,9 +1,0 @@
-package com.example.witt.data.source.remote.plan.detail_plan
-
-import com.example.witt.data.model.plan.detail_plan.request.AddPlaceRequest
-import com.example.witt.data.model.plan.detail_plan.response.AddPlaceResponse
-
-interface AddPlaceDataSource {
-
-    suspend fun addPlace(tripId: Int, dayId: Int, request: AddPlaceRequest): Result<AddPlaceResponse>
-}
