+++ client/app/src/main/java/com/example/witt/domain/model/remote/detail_plan/memo/EditMemoModel.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.plan.memo
+package com.example.witt.domain.model.remote.detail_plan.memo
 
 data class EditMemoModel (
     val status: Boolean,
