+++ client/app/src/main/java/com/example/witt/domain/model/remote/detail_plan/search/flight/AddFlightModel.kt
@@ -1,10 +1,10 @@
-package com.example.witt.domain.model.flight
+package com.example.witt.domain.model.remote.detail_plan.search.flight
 
 data class AddFlightModel(
     val status:Boolean,
     val reason:String,
     val planId:Int,
-    val planFlight:PlanFlight
+    val planFlight: PlanFlight
 )
 
 data class PlanFlight(
