+++ client/app/src/main/java/com/example/witt/domain/model/remote/detail_plan/search/flight/AddFlightRequest.kt
@@ -1,4 +1,4 @@
-package com.example.witt.domain.model.flight
+package com.example.witt.domain.model.remote.detail_plan.search.flight
 
 data class AddFlightRequest (
     val AirlineCode :String,
