+++ client/app/src/main/java/com/example/witt/presentation/widget/AddFlightDialog.kt
@@ -7,13 +7,13 @@ import android.util.Log
 import android.view.WindowManager
 import com.example.witt.data.repository.FlightRepositoryImpl
 import com.example.witt.databinding.DialogSearchFlightBinding
-import com.example.witt.domain.model.flight.SearchFlightModel
-import com.example.witt.domain.model.flight.toAddFlightRequest
+import com.example.witt.domain.model.remote.detail_plan.search.SearchFlightModel
+import com.example.witt.domain.model.remote.detail_plan.search.toAddFlightRequest
 import kotlinx.coroutines.*
 
 class AddFlightDialog(
     context: Context,
-    private val result:SearchFlightModel,
+    private val result: SearchFlightModel,
     private val tripId:Int,
     private val onClickApprove: () -> Unit = {},
     private val onClickCancel: () -> Unit = {}
@@ -46,7 +46,7 @@ class AddFlightDialog(
         )
     }
 
-    private fun initText(result:SearchFlightModel){
+    private fun initText(result: SearchFlightModel){
         binding.departureAirline.text = result.Flight.flyFrom
         binding.ariveAirline.text = result.Flight.flyTo
         binding.flightNum.text = result.Flight.flightNo.toString()
