+++ client/app/src/main/res/layout/item_plan_place.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="item"
-            type="com.example.witt.domain.model.plan.get_plan.PlanPlaceModel" />
+            type="com.example.witt.domain.model.remote.plan.get_plan.PlanPlaceModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
